+++ src/main/java/org/depromeet/sambad/moyeo/common/config/QuerydslConfig.java
@@ -1,0 +1,23 @@
+package org.depromeet.sambad.moyeo.common.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+import jakarta.persistence.EntityManager;
+
+@Configuration
+public class QuerydslConfig {
+
+	private final EntityManager entityManager;
+
+	public QuerydslConfig(EntityManager entityManager) {
+		this.entityManager = entityManager;
+	}
+
+	@Bean
+	public JPAQueryFactory jpaQueryFactory() {
+		return new JPAQueryFactory(entityManager);
+	}
+}
