--- src/main/java/org/depromeet/sambad/moyeo/common/infrastructure/QuerydslConfig.java
@@ -1,21 +1,0 @@
-package org.depromeet.sambad.moyeo.common.infrastructure;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class QuerydslConfig {
-
-    private final EntityManager entityManager;
-
-    public QuerydslConfig(EntityManager entityManager) {
-        this.entityManager = entityManager;
-    }
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-}
